bag_vetscan/make_analyzer_dracula/analyzer_app/tests/test_configuration_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.inline_response400 import InlineResponse400


async def test_configuration_factory_reset_put(client):
    """Test case for configuration_factory_reset_put

    
    """
    headers = { 
    }
    response = await client.request(
        method='PUT',
        path='/configuration/factory_reset',
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


async def test_configuration_get(client):
    """Test case for configuration_get

    
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/configuration',
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


async def test_configuration_put(client):
    """Test case for configuration_put

    
    """
    body = None
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
    }
    response = await client.request(
        method='PUT',
        path='/configuration',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


async def test_configuration_schema_get(client):
    """Test case for configuration_schema_get

    
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/configuration/schema',
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

